Added

+++ apstruct.cpp
@@ -1,10 +1,9 @@
 #include <cmath>
-#include <stdexcept>
-#include <sstream>
-
 #include "ap.h"
 
+
 using namespace std;
+
 
 // Round up to the nearest power of two or three times a power of two
 size_t rnd23up (size_t x)
@@ -650,7 +649,7 @@ apstruct *apmulmedium (apstruct *a, apstruct *b)
     size_t t, s, p, l, prec, size, ssize;
     apstruct *ap;
     modint *f, *src, *dest;
-    rawtype carry = 0, tmpdest[USE_N2_MUL];
+    rawtype carry = 0.0 , tmpdest[USE_N2_MUL];
     modint tmpf[USE_N2_MUL];
 
     assert (a);                 // Won't work on uninitialized apfloats
@@ -840,7 +839,7 @@ apstruct *apdivshort (apstruct *a, apstruct *b)
     {
         // Check for finite or infinite result sequence
         tmp1[0] = f;
-        for (t = 0; static_cast<int>(t) < NBasefactors; t++)
+        for (t = 0; t < NBasefactors; t++)
             while (bigdiv (tmp2, tmp1, Basefactors[t], 1) == 0) tmp1[0] = tmp2[0];
         if (tmp1[0] != 1)
         {
@@ -1099,7 +1098,7 @@ apstruct *apabsceil (apstruct *a)
     }
 
     // Check if the fractional part is nonzero
-    if (static_cast<long>(a->size) > a->exp)
+    if (a->size > a->exp)
         carry = 1;
     else
         carry = 0;
